# perf : Increase refresh threat min delay and optim on code