### Brief description of the changes:

Ragnarok77 ran some profiling and found that ~9% of the time spent in lategame threat farming was spent on updating the gui threat values. This attempts to solve that by precomputing the threat text, and applying it to every gui, rather than computing the threat text for every gui individually, and lengthening the time between threat updates (from every 5 ticks, to now every 30 ticks)


This PR is an alternative to https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/pull/393 where we removed subsecond threat updates entirely.  I have a suspicion the creation of the threat text was the majority of the runtime but I cannot be sure.

```
===
Last thing, here is a the summary of late game threat farmer lag : 
194694x "anonymous" in "...ing/Factorio/temp/currently-playing/utils/event_core.lua", line 57. Total Duration: 12065.984900ms

Out of 12k :

190645x "anonymous" in ".../Factorio/temp/currently-playing/functions/boss_unit.lua", line 31. Total Duration: 3269.118600ms [the joy of on damage event, it might be very fast to calculate, but it's called 200k times which is a lot...]

7x "refresh" in "...rio/temp/currently-playing/maps/biter_battles_v2/gui.lua", line 295. Total Duration: 900.294900ms [refresh the gui, bad gui optim]

470x "refresh_threat" in "...rio/temp/currently-playing/maps/biter_battles_v2/gui.lua", line 304. Total Duration: 1135.801400ms [just to refresh display threat :lulw: ]

                    16x C function "clone_area". Total Duration: 500.359500ms
                16x "clone" in "...rrently-playing/maps/biter_battles_v2/mirror_terrain.lua", line 31. Total Duration: 510.973600ms

                    8x "draw_biter_area" in "...temp/currently-playing/maps/biter_battles_v2/terrain.lua", line 316. Total Duration: 1587.792500ms (which is from GetNoise, get_noise.lua line 127) 
```

### Tested Changes:
- [ ] I've tested the changes locally or with people.
- [ ] I've not tested the changes.