@Ragnarok77-factorio 2023-10-18
Just for the trace , this show pretty is in loop, and redid every time, whereas it's the same, before this PR that fixes it .

In theory, just by moving this refresh pretty outside of loop, we should probably win like 300ms or something : 
![image](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/95619848/12b777d1-10fc-417c-9ce3-3df7915f545e)