@Ragnarok77-factorio 2023-10-18
Tested it, no issue found, I guess we try 394 first, then 393 otherwise if not good enough, will need profiler during big games and threat farmers to see if it improves anything or not