@Ragnarok77-factorio 2023-10-22
It worked well, now show pretty threat used only 22ms : 
![image](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/95619848/5f1f180e-20ab-45f1-a630-344eab58b9a5)